+++ internal/cli/healthcheck/command.go
@@ -5,9 +5,9 @@ import (
 	"fmt"
 	"strconv"
 
-	"github.com/modem7/docker-error-pages/internal/env"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
+	"github.com/modem7/docker-error-pages/internal/env"
 )
 
 type checker interface {